Added external command runner

+++ index.js
@@ -4,10 +4,11 @@ const git = require("./lib/git");
 const pkg = require("./lib/package");
 const cmd = require("./lib/commandline");
 const bumpVersion = require("semver-increment");
-const config = require("./lib/config");
+const config = require("./lib/config")();
 const utils = require("./lib/utils");
 const publishFrom = config.publishFrom || "master";
 const pushDevelBranch = config.pushDevelBranch === undefined ? true : config.pushDevelBranch;
+const commands = config.commands || [];
 
 const fs = require("fs");
 
@@ -35,32 +36,39 @@ if (cmd.check()) {
                     if (git.add_all()) {
                         // commit the changes
                         if (git.commit(`Patched files to ${newVersion}`)) {
-                            // merge the release branch to the branch that is being published from
-                            if (git.merge_to(publishFrom, relBranch)) {
-                                // tag the master branch
-                                if (git.tag_branch(newVersion)) {
-                                    // push the master branch
-                                    if (git.push_branch(publishFrom)) {
-                                        // delete the release branch
-                                        if (git.delete_branch(relBranch)) {
-                                            // switch to the dev branch
-                                            if (git.switch_branch(curBranch)) {
-                                                // set the version of the devel branch
-                                                if (
-                                                    utils.set_package_version(pkg.get_package_file(), newVersion, true)
-                                                ) {
-                                                    // push the devel branch
-                                                    if (pushDevelBranch) {
-                                                        if (git.push_branch(curBranch)) {
+                            // run external command from config
+                            if (utils.run_commands(commands)) {
+                                // merge the release branch to the branch that is being published from
+                                if (git.merge_to(publishFrom, relBranch)) {
+                                    // tag the master branch
+                                    if (git.tag_branch(newVersion)) {
+                                        // push the master branch
+                                        if (git.push_branch(publishFrom)) {
+                                            // delete the release branch
+                                            if (git.delete_branch(relBranch)) {
+                                                // switch to the dev branch
+                                                if (git.switch_branch(curBranch)) {
+                                                    // set the version of the devel branch
+                                                    if (
+                                                        utils.set_package_version(
+                                                            pkg.get_package_file(),
+                                                            newVersion,
+                                                            true
+                                                        )
+                                                    ) {
+                                                        // push the devel branch
+                                                        if (pushDevelBranch) {
+                                                            if (git.push_branch(curBranch)) {
+                                                                logger.info("All Done");
+                                                            }
+                                                        } else {
                                                             logger.info("All Done");
                                                         }
-                                                    } else {
-                                                        logger.info("All Done");
                                                     }
                                                 }
+                                            } else {
+                                                logger.error(`Deleting branch ${relBranch} failed!`);
                                             }
-                                        } else {
-                                            logger.error(`Deleting branch ${relBranch} failed!`);
                                         }
                                     }
                                 }
